SVC_Coins refactored + endpoints made idempotent

+++ src/SVC_Microservices/SVC_Coins/Repositories/CoinsRepository.cs
@@ -1,3 +1,4 @@
+using FluentResults;
 using Microsoft.EntityFrameworkCore;
 using SVC_Coins.Models.Entities;
 using SVC_Coins.Models.Input;
@@ -15,44 +16,72 @@ public class CoinsRepository(CoinsDbContext context) : ICoinsRepository
     private readonly CoinsDbContext _context = context;
 
     /// <inheritdoc />
-    public async Task InsertCoin(CoinNew coin)
+    public async Task<Result> InsertCoin(CoinNew coin)
     {
+        if (await CheckCoinExists(coin))
+            return Result.Fail("Coin already exists in the database.");
+
         var coinEntity = Mapping.ToCoinEntity(coin);
         await _context.Coins.AddAsync(coinEntity);
         await _context.SaveChangesAsync();
+        return Result.Ok();
     }
+
+    private async Task<bool> CheckCoinExists(CoinNew coin) =>
+        await _context.Coins.AnyAsync(c => c.Name == coin.Name && c.Symbol == coin.Symbol);
 
     /// <inheritdoc />
     public async Task<IEnumerable<Coin>> GetAllCoins()
     {
         var coinsWithTradingPairs = await _context.Coins.Include(c => c.TradingPairs).ToListAsync();
 
-        var coins = coinsWithTradingPairs.Select(Mapping.ToCoin);
-        return coins;
+        return coinsWithTradingPairs.Select(Mapping.ToCoin);
     }
 
     /// <inheritdoc />
-    public async Task DeleteCoin(int idCoin)
+    public async Task<Result> DeleteCoin(int idCoin)
     {
+        var coinToDelete = _context.Coins.Where(coin => coin.Id == idCoin);
+        if (!coinToDelete.Any())
+            return Result.Fail($"Coin with ID {idCoin} not found.");
+
         var tradingPairsToDelete = _context.TradingPairs.Where(tp =>
             tp.IdCoinMain == idCoin || tp.IdCoinQuote == idCoin
         );
         _context.TradingPairs.RemoveRange(tradingPairsToDelete);
 
-        var coinToDelete = _context.Coins.Where(coin => coin.Id == idCoin);
         _context.Coins.RemoveRange(coinToDelete);
 
         await _context.SaveChangesAsync();
+        return Result.Ok();
     }
 
     /// <inheritdoc />
-    public async Task<int> InsertTradingPair(TradingPairNew tradingPair)
+    public async Task<Result<int>> InsertTradingPair(TradingPairNew tradingPair)
     {
         var tradingPairEntity = Mapping.ToTradingPairEntity(tradingPair);
+
+        if (!await CheckCoinsExist(tradingPairEntity))
+            return Result.Fail("One or both coins do not exist in the Coins table.");
+
+        if (await CheckTradingPairExists(tradingPairEntity))
+            return Result.Fail("This trading pair already exists.");
+
         await _context.TradingPairs.AddAsync(tradingPairEntity);
         await _context.SaveChangesAsync();
-        return tradingPairEntity.Id;
+
+        return Result.Ok(tradingPairEntity.Id);
     }
+
+    private async Task<bool> CheckCoinsExist(TradingPairEntity tradingPairEntity) =>
+        await _context.Coins.AnyAsync(coin => coin.Id == tradingPairEntity.IdCoinMain)
+        && await _context.Coins.AnyAsync(coin => coin.Id == tradingPairEntity.IdCoinQuote);
+
+    private async Task<bool> CheckTradingPairExists(TradingPairEntity tradingPairEntity) =>
+        await _context.TradingPairs.AnyAsync(tp =>
+            tp.IdCoinMain == tradingPairEntity.IdCoinMain
+            && tp.IdCoinQuote == tradingPairEntity.IdCoinQuote
+        );
 
     /// <inheritdoc />
     public async Task<IEnumerable<Coin>> GetQuoteCoinsPrioritized()
